# pet-services-api/internal/transport/http/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is an API for managing restaurants.",
        "title": "Pedido Simples API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Pedido Simples",
            "url": "http://www.pedidosimples.com.br",
            "email": "contato@pedidosimples.com.br"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.LoginOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Logout",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003crefresh_token\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/refresh": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh tokens",
                "parameters": [
                    {
                        "description": "Refresh payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.refreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.RefreshOutput"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/signup": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Signup",
                "parameters": [
                    {
                        "description": "Signup payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.signupRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/auth.SignupOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "auth.LoginOutput": {
            "type": "object",
            "properties": {
                "accessExpiresAt": {
                    "type": "integer"
                },
                "accessToken": {
                    "type": "string"
                },
                "refreshExpiresAt": {
                    "type": "integer"
                },
                "refreshToken": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                },
                "userType": {
                    "$ref": "#/definitions/user.UserType"
                }
            }
        },
        "auth.RefreshOutput": {
            "type": "object",
            "properties": {
                "accessExpiresAt": {
                    "type": "integer"
                },
                "accessToken": {
                    "type": "string"
                },
                "refreshExpiresAt": {
                    "type": "integer"
                },
                "refreshToken": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                },
                "userType": {
                    "type": "string"
                }
            }
        },
        "auth.SignupOutput": {
            "type": "object",
            "properties": {
                "accessExpiresAt": {
                    "type": "integer"
                },
                "accessToken": {
                    "type": "string"
                },
                "refreshExpiresAt": {
                    "type": "integer"
                },
                "refreshToken": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                },
                "userType": {
                    "$ref": "#/definitions/user.UserType"
                }
            }
        },
        "exceptions.ProblemDetails": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "instance": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "exceptions.ProblemDetailsOutputDTO": {
            "type": "object",
            "properties": {
                "problem_details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/exceptions.ProblemDetails"
                    }
                }
            }
        },
...
        "http.loginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
        "http.refreshRequest": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
...
        "http.signupRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password",
                "phone",
                "type"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "password": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 8
                },
                "phone": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 10
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "owner",
                        "provider"
                    ]
                }
            }
        },
...
        "user.UserType": {
            "type": "string",
            "enum": [
                "owner",
                "provider"
            ],
            "x-enum-varnames": [
                "UserTypeOwner",
                "UserTypeProvider"
            ]
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}